Add login HTML file to repository.
This is an HTML login form that allows users to enter their username and password to access a website or application. The form contains two input fields, one for the username and one for the password, along with a “Remember me” checkbox and a “Login” button. The form also includes a “Forgot password?” link that redirects users to another page where they can reset their password if they have forgotten it.